Add new endpoints, update SeedRoles, and modify DB schema
- Added `AddImage` endpoint in `CategoryController` to add images to categories.
- Commented out `ConfirmWebhook` and `HandlePayOSTransfer` in `PaymentController`.
- Updated `SeedRoles` to create a wallet for the admin user.
- Changed `DefaultConnection` strings in `appsettings` files to point to `BlindBoxV3`.
- Removed `20250319092313_init.Designer.cs` and `20250319092313_init.cs` migration files.
- Added new properties and relationships to various models and DTOs.
- Updated `Wallet` and `WalletTransaction` models to use `decimal` for financial fields.
- Introduced new migration files to reflect schema changes.

+++ Services/Wallet/WalletTransactionService.cs
@@ -17,7 +17,7 @@ namespace Services.Wallet
             _walletTransactionRepository = walletTransactionRepository;
         }
 
-        public async Task AddWalletTransactionAsync(Guid walletId, int amount, string transactionType, string transactionStatus, string transactionDate, int transacionBalance, int? orderId)
+        public async Task AddWalletTransactionAsync(Guid walletId, decimal amount, string transactionType, string transactionStatus, string transactionDate, decimal transacionBalance, int? orderId)
         {
             WalletTransaction walletTransaction = new WalletTransaction
             {
